Enemy Health and Improved Hit Effect
Implement concept of Enemy Health. Detect hits on the Zombie's limbs and take damage. Destroy Zombie when health reaches zero. Provoke Zombie into chasing player if a hit is detected from outside its provocation ranges. Randomize the rotation on the hit effect for a nicer user experience, especially when the player shoots straight on as the effect is currently a 2D object.